Intersections folder.
